seperate app from index
